#102 Added the custom view to it's own separate submodule
