+++ list/emptyRecyclerView/src/main/java/be/appwise/emptyRecyclerView/RecyclerViewState.kt
@@ -1,0 +1,7 @@
+package be.appwise.emptyRecyclerView
+
+enum class RecyclerViewState {
+    LOADING,
+    NORMAL,
+    EMPTY
+}
